Fixed It. It works disclaimer:
CONTENT WARNING MAY CONTAIN BAD CODE

So it works... to a point. the rover initial start point cant have a fixed angle so we assume it is sitting able to move forward. However we can set an approach angle for the target node so that we can make sure it is going to probe the correct side.